backend/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-cors-middleware==1.5.0
django-oauth-toolkit==1.3.2
djangorestframework==3.11.1
idna==2.10
oauthlib==3.1.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
